Start initial abstraction
This creates basic abstractions for languages and repo servers.
The goal long term would be to support multiple source repos like
github, gitlab, etc. as well as multiple language package
ecosystems like Go, Rust, Java, etc.
