+++ src/lib.rs
@@ -13,19 +13,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod bundle;
 pub mod models;
 
+use bundle::bundle::{EcosystemInitConfig, GithubBundle};
 use inquire::Text;
-use models::security_insights::insights10::{
-    SecurityInsightsVersion100YamlSchema, SecurityInsightsVersion100YamlSchemaContributionPolicy,
-    SecurityInsightsVersion100YamlSchemaHeader,
-    SecurityInsightsVersion100YamlSchemaHeaderSchemaVersion,
-    SecurityInsightsVersion100YamlSchemaProjectLifecycle,
-    SecurityInsightsVersion100YamlSchemaProjectLifecycleStatus,
-    SecurityInsightsVersion100YamlSchemaVulnerabilityReporting,
-};
 use octocrab::Page;
-use std::{error::Error, fs, process::Command, thread, time::Duration};
+use std::error::Error;
 
 pub async fn create() -> std::result::Result<(), Box<dyn Error>> {
     let instance = octocrab::instance();
@@ -47,180 +41,36 @@ pub async fn create() -> std::result::Result<(), Box<dyn Error>> {
     )
     .prompt()?;
 
-    create_repository(&instance, &name, &description, &organization).await?;
-    create_readme(&instance, &name, &description, &organization).await?;
-    clone_repository(&organization, &name).await?;
-    initialize_go_module(&organization, &name).await?;
-    create_security_insights_yaml(&organization, &name).await?;
-    commit_and_push_changes(&name).await?;
-
-    Ok(println!("Created a new skootrs project"))
-}
-
-async fn create_repository(
-    instance: &octocrab::Octocrab,
-    name: &str,
-    description: &str,
-    organization: &str,
-) -> std::result::Result<(), Box<dyn Error>> {
-    instance
-        .repos("kusaridev", "skoot-go")
-        .generate(name)
-        .owner(organization)
-        .description(description)
-        .send()
-        .await?;
-    println!(
-        "Created {} repository from template: kusaridev/skoot-go",
-        name
+    let language = inquire::Select::new(
+        "Select a language",
+        bundle::bundle::SUPPORTED_ECOSYSTEMS.to_vec(),
     );
-    // TODO: figure out what the best way to wait for the repository to be created and template generated
-    thread::sleep(Duration::from_millis(4000));
-    Ok(())
-}
 
-async fn create_readme(
-    instance: &octocrab::Octocrab,
-    name: &str,
-    description: &str,
-    organization: &str,
-) -> std::result::Result<(), Box<dyn Error>> {
-    instance
-        .repos(organization, name)
-        .create_file(
-            "README.md",
-            "Create README",
-            format!(
-                r#"# {}
-{}
-"#,
-                name, description
-            ),
-        )
-        .send()
-        .await?;
-    println!("Created README.md for {}", name);
-    Ok(())
-}
-
-async fn clone_repository(
-    organization: &str,
-    name: &str,
-) -> std::result::Result<(), Box<dyn Error>> {
-    let url = format!("https://github.com/{}/{}", organization, name);
-    let _output = Command::new("git")
-        .arg("clone")
-        .arg(&url)
-        .arg(format!("/tmp/{}", name))
-        .output()
-        .expect("Failed to execute git clone command");
-    println!("Cloned {} to /tmp/{}", url, name);
-    Ok(())
-}
-
-async fn initialize_go_module(
-    organization: &str,
-    name: &str,
-) -> std::result::Result<(), Box<dyn Error>> {
-    let _output = Command::new("go")
-        .arg("mod")
-        .arg("init")
-        .arg(format!("github.com/{}/{}", organization, name))
-        .current_dir(format!("/tmp/{}", name))
-        .output()
-        .expect("Failed to execute go mod init command");
-    println!("Initialized go module for {}", name);
-    Ok(())
-}
-
-async fn create_security_insights_yaml(
-    organization: &str,
-    name: &str,
-) -> std::result::Result<(), Box<dyn Error>> {
-    let url = format!("https://github.com/{}/{}", organization, name);
-    fs::write(
-        format!("/tmp/{}/SECURITY_INSIGHTS.yaml", name),
-        create_security_insights(url)?,
-    )?;
-    println!("Created SECURITY_INSIGHTS.yaml for {}", name);
-    Ok(())
-}
-
-async fn commit_and_push_changes(name: &str) -> std::result::Result<(), Box<dyn Error>> {
-    let _output = Command::new("git")
-        .arg("add")
-        .arg(".")
-        .current_dir(format!("/tmp/{}", name))
-        .output()
-        .expect("Failed to execute git add command");
-
-    let commit_message = format!("Initialize go module for {}", name);
-    let _output = Command::new("git")
-        .arg("commit")
-        .arg("-m")
-        .arg(&commit_message)
-        .current_dir(format!("/tmp/{}", name))
-        .output()
-        .expect("Failed to execute git commit command");
-    println!("Committed changes to {}", name);
-
-    let _output = Command::new("git")
-        .arg("push")
-        .current_dir(format!("/tmp/{}", name))
-        .output()
-        .expect("Failed to execute git push command");
-    println!("Pushed changes to {}", name);
-    Ok(())
-}
-
-pub fn create_security_insights(url: String) -> std::result::Result<String, Box<dyn Error>> {
-    let insights = SecurityInsightsVersion100YamlSchema {
-        contribution_policy: SecurityInsightsVersion100YamlSchemaContributionPolicy {
-            accepts_automated_pull_requests: true,
-            accepts_pull_requests: true,
-            automated_tools_list: None,
-            code_of_conduct: None,
-            contributing_policy: None,
-        },
-        dependencies: None,
-        distribution_points: Vec::new(),
-        documentation: None,
-        header: SecurityInsightsVersion100YamlSchemaHeader {
-            changelog: None,
-            commit_hash: None,
-            expiration_date: chrono::Utc::now() + chrono::Duration::days(365),
-            last_reviewed: None,
-            last_updated: None,
-            license: Some(format!("{}/blob/main/LICENSE", &url)),
-            project_release: None,
-            project_url: url,
-            schema_version: SecurityInsightsVersion100YamlSchemaHeaderSchemaVersion::_100,
-        },
-        project_lifecycle: SecurityInsightsVersion100YamlSchemaProjectLifecycle {
-            bug_fixes_only: false,
-            core_maintainers: None,
-            release_cycle: None,
-            release_process: None,
-            roadmap: None,
-            status: SecurityInsightsVersion100YamlSchemaProjectLifecycleStatus::Active,
-        },
-        security_artifacts: None,
-        security_assessments: None,
-        security_contacts: Vec::new(),
-        security_testing: Vec::new(),
-        vulnerability_reporting: SecurityInsightsVersion100YamlSchemaVulnerabilityReporting {
-            accepts_vulnerability_reports: false,
-            bug_bounty_available: None,
-            bug_bounty_url: None,
-            comment: None,
-            email_contact: None,
-            in_scope: None,
-            out_scope: None,
-            pgp_key: None,
-            security_policy: None,
-        },
+    let ecosystem_init_config: EcosystemInitConfig = match language.prompt()? {
+        "Go" => {
+            let go_config = bundle::bundle::GoConfig {
+                // TODO: Support more than Github
+                module: format!("github.com/{}/{}", organization, name),
+            };
+            EcosystemInitConfig::Go(go_config)
+        }
+        "Maven" => {
+            let maven_config = bundle::bundle::MavenConfig {};
+            EcosystemInitConfig::Maven(maven_config)
+        }
+        _ => {
+            unreachable!("Unsupported language");
+        }
     };
 
-    let yaml = serde_yaml::to_string(&insights)?;
-    Ok(yaml)
+    let bundle = GithubBundle::new(
+        &instance,
+        &name,
+        &description,
+        &organization,
+        ecosystem_init_config,
+    );
+    bundle.create().await?;
+
+    Ok(println!("Created a new skootrs project"))
 }
